refactor(web): workflow constant code splitting
