+++ web/src/views/Workflow/components/Properties/index.tsx
@@ -2,7 +2,7 @@
  * @Author: ZhaoYing 
  * @Date: 2026-02-03 15:39:59 
  * @Last Modified by: ZhaoYing
- * @Last Modified time: 2026-06-30 11:59:12
+ * @Last Modified time: 2026-07-15 11:05:08
  */
 import { type FC, useEffect, useState, useMemo } from "react";
 import clsx from 'clsx'
@@ -35,7 +35,7 @@ import Editor, { type LexicalEditorProps } from "../Editor";
 import RbSlider from '@/components/RbSlider'
 import JinjaRender from './JinjaRender'
 import CodeExecution from './CodeExecution'
-import { nodeLibrary } from '../../constant';
+import { nodeLibrary, cannotRunNodes } from '../../constant';
 import RbCard from '@/components/RbCard/Card';
 import ModelConfig from './ModelConfig'
 import ModelSelect from '@/components/ModelSelect'
@@ -43,7 +43,6 @@ import ListOperator from './ListOperator'
 import MappingList from "./MappingList";
 import ErrorHandle from './ErrorHandle'
 import SingleNodeRun from '../SingleNodeRun'
-import { cannotRunNodes } from '../../constant'
 import RadioGroupBtn from './RadioGroupBtn'
 import type { Model } from '@/views/ModelManagement/types';
 import Retry from './Retry'
