+++ web/src/views/Workflow/constant/index.ts
@@ -1,0 +1,17 @@
+/*
+ * @Author: ZhaoYing
+ * @Date: 2026-02-03 15:06:18
+ * @Last Modified by: ZhaoYing
+ * @Last Modified time: 2026-07-15 10:53:26
+ *
+ * 说明：原单文件（约 1491 行）已按主题拆分至 ./constants/ 目录，
+ * 此处作为桶文件（barrel）重新导出，保持所有对外引用路径与符号不变。
+ */
+export * from './nodeFlags';
+export * from './nodeLibrary';
+export * from './theme';
+export * from './layout';
+export * from './nodeShapes';
+export * from './ports';
+export * from './graphNodeLibrary';
+export * from './edges';
