+++ web/src/views/Workflow/constant/layout.ts
@@ -1,0 +1,6 @@
+/* 布局尺寸常量：节点宽度与条件节点高度相关（拆分自 constant.ts） */
+export const nodeWidth = 240;
+
+export const conditionNodePortItemArgsY = 56.5;
+export const conditionNodeItemHeight = 26;
+export const conditionNodeHeight = 110;
